AI Controller implemented for A5. GUI bug fixed.
Now the game can be played between any two of the human player, dumb AI
and random AI by first select the player for X and O, then press the
start button.
The GUI bug is fixed by commenting two getters of x and y in the Square
class out. NO REASON why these two simple getters that NEVER BEEN USED
will cause the bug of displaying the GUI. WEIRD.
Another bug of the program is function repaint() is optimized by the
complier to be executed after all other functions of the program
returned, no matter where it is written. CANNOT CONTROL when it is
executed!
So when the game is played by two AIs, the GUI won’t be updated before
the game is finished because repaint() won’t be executed until then.
This SUCKS. GUI classes SUCK.